Add files via upload
Kandeez is a kind of ransomware worm that encrypts many of your documents, photos, videos, databases, and other important files, making them accessible because they have been encrypted, making it impossible for users to access their files without a decryption tool and valid key. This tool is developed to help Red Teams and ethical hackers simulate a real ransomware attack.

There is no shortcut for something like this.

This is a ransomware simulator that can safely encrypt some files on your computer to test whether your defences are actually working.